go.text/locale: Separated Macro canonicalization into Legacy and Macro groups, as defined by CLDR. The Legacy cases are now hard coded. This allows us to handle sh -> sr-Latn without introducing a new data type just for this case. The set of legacy translations is unlikely to change, but maketables.go now checks and fails if the set changes. Also introduced Default CanonType in preperation for adding tag maximization and minimization.
Further changes: deviating from CLDR in a few places to not have to deal with
legacy choice. CLDR is likely to head in this direction as well, so it
prevents incompatibilities down the road.
Added CLDR option to force strict compliance to CLDR.

Mapping "mo" to "ro-MD" instead of "ro".  In cases where ID is used as a
locale, preserving this piece of information may be important. It is up
to the matching code to establish that "ro" and "ro-MD" are mutually
intelligible.

R=r
CC=golang-dev
https://golang.org/cl/12903045