go.text/language: added predefined common tag values.
The benefits of using predefined tags are:
- convenience for the user
- an additional layer or indirection for tag representations that may
  change in the future (e.g. pt to mean Brazilian Portuguese or just
  Portuguese).

Predefined values are only provided for a selection of tags. The selection
is based on CLDR data and augmented with languages for populous areas where
speakers do not typically also fluently speak a language that was already
included in the set (thereby increasing the likelihood of needing it) and
languages for which tags may be ambiguous.

We do not intend to provide similar constants for Regions, Scripts and Base
values. Specifically the set of Regions is much more "dynamic" than the set
of languages. The (Tag|Base|Script|Region|Currency)OrDie methods are provided
as a convenience for the user to specify values at startup.

Details:
- Went with names instead of codes to allow for a level of indirection on
  top of tags.
- Names for the tags are taken from CLDR for the en locale.
- Internal tag constants are now prefixed with '_' instead of lang_, reg,
  scr and cur. All tags are unique between these for so this does not pose
  a problem. It makes the predefined values for tags look a lot better in
  the godocs.

LGTM=r
R=r
CC=golang-codereviews
https://golang.org/cl/59740047