go.text/language: fixed bug in Script, where sometimes the wrong likely script is returned.
The algorithm previously assumed that if the IANA registry has a
SuppressScript entry for a language, this is typically the only script in
use for this language. There are a few langauges for which this is not the
case, though. For example, Kazakh is typically written in Arabic (and
possibly Latin) script by native speakers in China, instead of the Cyrillic
typically used in Kazakhstan. A similar situation exists for Panjabi and
Malay.

LGTM=r
R=r
CC=golang-codereviews
https://golang.org/cl/74680044