go.text/language: map grandfathered tags to ICU values, instead of CLDR.
The mapping is now hand-coded. This is fine as the table is small and
not likely to change.

ICU defines mappings for all grandfathered tags, whereas CLDR does not.
Furthermore, ICU maps zh-guoyu to cmn, instead of zh. cmn is slightly more
informative and arguably preferable. Note that the Matcher will map
cmn to zh without issue when appropriate. Also, canonicalizing with the
Macro option will map cmn to zh as well.

LGTM=roubert, r
R=roubert, r
CC=golang-codereviews
https://golang.org/cl/160790043