text/internal/colltab: copy contract*.go files
Files are copied from text/collate/colltab to text/internal/colltab.
This starts a new, internal colltab package. Collation is moving to fractional
weights. Almost everything in colltab needs to be rewritten as a result,
but the contraction trie can be reused.

The package is in text/internal, instead of text/collate/internal as other
packages may need to use collation tables (e.g. search).

Change-Id: I55fb3f439a4924e807bb301ad29f8b485d54bf1f
Reviewed-on: https://go-review.googlesource.com/4550
Reviewed-by: Nigel Tao <nigeltao@golang.org>
Reviewed-by: Rob Pike <r@golang.org>