runes: Added Transformer wrapper.
This brings this package in line with other packages in text and
makes it clearer that the functions in the package are transformers.

Note that the Bytes and String function check for errors. An error
will almost never occur, though. The only case where a Transformer
will result in an error is if the user passes a Transformer that may
generate an error to If.

Change-Id: I38d8bba4575f6b2955d4d687fc7a3a1fdd0f46b4
Reviewed-on: https://go-review.googlesource.com/9130
Reviewed-by: Nigel Tao <nigeltao@golang.org>