language: added Currencies to the Coverage type.
This is the first in a series of currency-related CLs.

Currencies have been not fully supported until now as I wasn't sure
whether to put them in a separate package or included them here.
There are several reasons to take the approach as originally
envisoned:
1) Currencies are somewhat looser, but still quite coupled with the
other types in Region.
2) It would be quite a small package on its own.
3) It allows piggybacking on existing types like Coverage.
4) There are not many other types that we can expect to add
   (the set of locale-related types has been quite steady in
   other systems).
5) Currencies is usually included in locale-related functionality
   in other libraries.
6) We would have to deal with some circular dependencies.
7) Moving Currency out is not backward compatible.

The main problem with keeping Currency is that language.Currency
is somewhat ugly. By hindsight, locale might still have been a
better name for this package. It is consistent with language.Region
and language.Script, though.

Change-Id: If7c4ebf5ca014881fe5866eaa9e132babdf0c2f6
Reviewed-on: https://go-review.googlesource.com/11637
Reviewed-by: Nigel Tao <nigeltao@golang.org>