unicode/rangetable: added assigned tables
See tables.go for a definition of assigned. It can be useful to have the
historical assigned code points to create programs where stability is
of importance.

Assigned is a function, instead of a map, to allow for future
optimizations. For example, the different versions can be stored
as diffs of each other and then have the various version be lazily
constructed.

Change-Id: I6ce8f3ab6f04edcf1b0ad7d2b98b21daee081f50
Reviewed-on: https://go-review.googlesource.com/11781
Reviewed-by: Nigel Tao <nigeltao@golang.org>