language: add all languages with data in compact set
maketables assigns a low code to languages that are frequent.
Include all languages for which CLDR defines data in this set.
This guarantees there is a compact index for all languages
references in gen_index.go. This, in turn, can be used to
optimize CompactIndex, which has some issues (explained in
later CL).

This CL also includes the addition of CLDRVersion and
deprecation of Version. This is to make it consistent
with other packages.
(Sorry, thought I did excluded it from staging)

Change-Id: Ia92e8c663fd4641976268c763645bcea2d1679a6
Reviewed-on: https://go-review.googlesource.com/16624
Reviewed-by: Hyang-Ah Hana Kim <hyangah@gmail.com>