internal/ucd: allow non-runes in expansion mode
The first field of UCD files are not always single runes or ranges.
Up till now one had to pass the KeepRanges option to prevent the
first field from being interpreted. Now it will ignore the field
setting a single-rune range with value 0. It also sets keepRanges
to force an error if the user tries to access the field as a Range.

Change-Id: Id3211cc94684cd37c5449c24e45de85d32687298
Reviewed-on: https://go-review.googlesource.com/17292
Reviewed-by: Hyang-Ah Hana Kim <hyangah@gmail.com>