encoding: encoders return errors on out-of-repertoire runes
This is the recommended behavior by WhatWG. The ErrorHandlers
introduced in another CL can be used to replace errors with,
for example, HTML escape sequences.

This CL also makes a first step with putting tests in the same
directory as the code tested, instead of putting all tests for
encodings in a single directory. This just works better for
a variety of things. Some other tests can be moved in a later CL.

Change-Id: I8f7f7e376f01b8994080488e8ebea2e15810aa4b
Reviewed-on: https://go-review.googlesource.com/17702
Reviewed-by: Andy Balholm <andy@balholm.com>
Reviewed-by: Hyang-Ah Hana Kim <hyangah@gmail.com>