: run long tests after a generated package
To build the repo one should typically use the UNICODE_DIR env var to
allow reasonable build times.

This change standardizes the long option from tests and causes long
tests to run if the UNICODE_DIR env var is set. It thereby ensures
that these tests are run when the repo is built.

Change-Id: I26f5656069146d4966d773eafd0a2e4797f18f22
Reviewed-on: https://go-review.googlesource.com/17963
Reviewed-by: Hyang-Ah Hana Kim <hyangah@gmail.com>