internal/format/plural: added plural package
This package will be moved out of internal at some point.

In the proposal (12750) this package is slated to be a
subpackage of feature. In practice format/feature/plural
makes more sense. But as there is no need for the feature
as of yet (or ever), we put it directly under format.

Change-Id: I8d5c9be00d458087680662dc6c117eee2dcfc566
Reviewed-on: https://go-review.googlesource.com/19333
Reviewed-by: Nigel Tao <nigeltao@golang.org>