unicode/bidi: implemented Properties type
API is analoguous to unicode/norm.Properties.

Unfortunately, the trie lookup code needed to be
copied and modified, mainly because of semantics,
but also because of performance.
Both issues could be addressed by adding options
to the trie generator.  A TODO is added for this.

Change-Id: I5a23d3bae39f2d7669fc390bdf5e72beb4bbba4d
Reviewed-on: https://go-review.googlesource.com/23563
Reviewed-by: Nigel Tao <nigeltao@golang.org>
Run-TryBot: Marcel van Lohuizen <mpvl@golang.org>
TryBot-Result: Gobot Gobot <gobot@golang.org>