secure/precis: API change
Profile is now a pointer. There is a bit too much stuff
in it for it to be passed by value. We could make it
smaller, but way we don't have to think about it,
especially since there seems to be no real benefit
to having it as value.

Also upgraded enforce tests to use subtests, which
makes testing a bit easier.

Change-Id: I678b8ae59ae3da2cf927f291aae6cc5e6c361da0
Reviewed-on: https://go-review.googlesource.com/23595
Reviewed-by: Nigel Tao <nigeltao@golang.org>
Run-TryBot: Marcel van Lohuizen <mpvl@golang.org>