internal/number: copy of strconv formatting code
Copied code of strconv instead of fmt, as the plan
is to add a low-level API to strconv to support some
of the functionality needed here and delete this code.

Public API will be removed in a follow-up CL. This is
okay as this is an internal package.

Change-Id: Ia9dc81fb537fd96e31417878b141219a048f31e5
Reviewed-on: https://go-review.googlesource.com/25131
Run-TryBot: Marcel van Lohuizen <mpvl@golang.org>
Reviewed-by: Nigel Tao <nigeltao@golang.org>