cases: “fixed” most of the ICU conformance issues
As it turns out, by far most issues were a problem of ICU,
not our implementation. ICU has (at least) two implementations
for casing and it is confirmed that the non-conformance issues that
are dismissed at least match the results of the other implementation.

The non-conformance with the long string of modifiers is a result
of our approach of making Unicode stream-safe.

Change-Id: I8e9fec1fc51d38a1950f95be72e07fb80d9949e2
Reviewed-on: https://go-review.googlesource.com/29183
Reviewed-by: Nigel Tao <nigeltao@golang.org>
Run-TryBot: Marcel van Lohuizen <mpvl@golang.org>
TryBot-Result: Gobot Gobot <gobot@golang.org>