message/catalog: added new package
This package is based on functionality extracted from the
message package. It adds supports for variables and macros
in messages.

Note that the documentation refers to packages that are
currently under construction, but this is intentionally kept
in to show where this package will be used for.

Note that the way Messages are represented is rather
inefficient. This package has been designed, though, to
allow for an efficient representation later on.

Change-Id: I846bc930dabb296beb156b482cde55d4eeb2c2a1
Reviewed-on: https://go-review.googlesource.com/38694
Run-TryBot: Marcel van Lohuizen <mpvl@golang.org>
TryBot-Result: Gobot Gobot <gobot@golang.org>
Reviewed-by: Nigel Tao <nigeltao@golang.org>