language: simplify docs
Most use cases will be limited to the given examples.

It now refers to the blog for more details.

Removed section on rolling your own
locale-specific service as it added too much overhead
to the doc, should be rare and also the example did not
apply to all cases.

Change-Id: I9db582c447d2a707ff83ee0752ef4ceb97306b2c
Reviewed-on: https://go-review.googlesource.com/59814
Run-TryBot: Marcel van Lohuizen <mpvl@golang.org>
Reviewed-by: Nigel Tao <nigeltao@golang.org>