internal/format: add interface
This is analogous to fmt.Formatter.
It is a bit questionable to use fmt.Formatter for
localized text rendering as it is unclear what the
meaning is in some cases. Use this until we know
for sure.

Change-Id: I3571c1bb3db1efeb6e7daf23afd7a6147b5a3dbc
Reviewed-on: https://go-review.googlesource.com/60731
Run-TryBot: Marcel van Lohuizen <mpvl@golang.org>
TryBot-Result: Gobot Gobot <gobot@golang.org>
Reviewed-by: Nigel Tao <nigeltao@golang.org>