x/text: Unicode 10 upgrade
This also includes an update of the IDNA
algorithm. Some of the definitions of the
algorithm changed along with the data,
so it is not possible to split this up in
separate CLs.

See Issue golang/go#21471.

The IDNA change is an update to the interpretation
of the RFC by UTS golang/go#46, which affects the bidi
rule.
Also the interpretation of empty labels has been
sharpened.

The following files have manual changes
(all related to IDNA):
- secure/precis/enforce_test.go
- secure/bidirule/bidirule.go
- secure/bidirule/bidirule_test.go
- internal/export/idna/idna.go
- internal/export/idna/idna_test.go

Change-Id: I4365787c3714b2fa392e7b8ae70b3b446637935c
Reviewed-on: https://go-review.googlesource.com/63791
Run-TryBot: Marcel van Lohuizen <mpvl@golang.org>
Reviewed-by: Nigel Tao <nigeltao@golang.org>