cmd/gotext: distinguish between key and id
The id of the translation and the key that is used at
runtime to look up the translation are not the same.
The id is used for tracking and linking messages.
The key is used for lookup internally.

Change-Id: I26cac35fb421ee8227fd8b0b5a712a752c4f5f5e
Reviewed-on: https://go-review.googlesource.com/82202
Run-TryBot: Marcel van Lohuizen <mpvl@golang.org>
TryBot-Result: Gobot Gobot <gobot@golang.org>
Reviewed-by: Nigel Tao <nigeltao@golang.org>