language: change semantics of CompactIndex
This is the first CL in a sequence to enhance both the
performance and functionality of language Tags.

First step is to have CompactIndex implement the
semantics that it should have: rather than returning
Und when there is no match, it should return the closest
matching compact index. CompactIndex is often used
this way, and when it is not it was typcially a bug. This
fixes that.

Next steps are to have the fast-path internal encoding
of tag be a compact index.

The step after that is to keep two compact indexes,
allow representing both the language and locale part
of a tag like:
    en-US-va-posix
and
    en-GB-rg-gbsct

The result will allow for dual-mode tags resulting from
matching, simplifying the life of an API user.
(In Java the API users will get two tags from a match
and will have to decide on a case-by-case basis which
tag to use for which purpose.)
This design also allows for fast converstion to compact
indices, making it unnecessary for message.Printers to
extract this index.

Change-Id: I019ac6df4a3e3c7e6c758173fc942667df65ff36
Reviewed-on: https://go-review.googlesource.com/95815
Run-TryBot: Marcel van Lohuizen <mpvl@golang.org>
TryBot-Result: Gobot Gobot <gobot@golang.org>
Reviewed-by: Nigel Tao <nigeltao@golang.org>