language/internal: finalize Builder and refine -u semantics
This is to aid the Compose implementation in language.

This also treats key-value pairs in -u extensions as individual
entities, allowing them to be merged and overwritten.

Multiple values of the same key are now disallowed; the
first value is taken. It is not considered an error, though,
allowing tags with different values to be composed.

Change-Id: I65f1e2a2645d15289a973356b9cae33d66068c2b
Reviewed-on: https://go-review.googlesource.com/95831
Run-TryBot: Marcel van Lohuizen <mpvl@golang.org>
Reviewed-by: Nigel Tao <nigeltao@golang.org>