internal/language/compact: copy of language
Copy language package to internal/language/compact
The purpose is to expose the compact ID functionality
to other packages in x/text, but without exporting
it outside of x/text.

The next steps will be to delete respective
functionality in the original and destination
package and expose the functionality in the
the internal package.

Change-Id: I35ff1a041069729f75bab743148f4cf3f16c2058
Reviewed-on: https://go-review.googlesource.com/96637
Run-TryBot: Marcel van Lohuizen <mpvl@golang.org>
TryBot-Result: Gobot Gobot <gobot@golang.org>
Reviewed-by: Ross Light <light@google.com>