encoding: regenerate tables
The default indentation for maps changed in a recent Go commit. Rerun
"go generate" for these packages to update the tables to match.

Also add the "replacement" key to the list of encodings retrievable
by Get. It is described in the W3C specification as "[existing] to
prevent certain attacks that abuse a mismatch between encodings
supported on the server and the client."

Change-Id: I3c671c229d1e68721d109364aaf19021bbd3af89
Reviewed-on: https://go-review.googlesource.com/127923
Reviewed-by: Marcel van Lohuizen <mpvl@golang.org>
Run-TryBot: Marcel van Lohuizen <mpvl@golang.org>
TryBot-Result: Gobot Gobot <gobot@golang.org>