language: fix TestBestMatchAlloc test
The test calls testing.AllocsPerRun with a count of 1.
AllocsPerRun sets runtime.GOMAXPROCS to 1, but it doesn't prevent the runtime from descheduling a goroutine
and performing an allocation somewhere in the background.
This commit changes the test to use a number of runs large enough to average away the occasional noisy allocation.

Fixes https://github.com/golang/go/issues/45809