go.mod: ignore cyclic dependency for tagging
tools and text have a cyclic dependency, which needs to be broken during
the upgrading/tagging process. This direction is due to
message/pipeline, which is marked as "IN DEVELOPMENT", so I think it's
less important to keep it up to date.

For golang/go#48523.

Change-Id: Ic4cdfd05628649035eee8a37e662cf22827c705b
Reviewed-on: https://go-review.googlesource.com/c/text/+/439295
Run-TryBot: Heschi Kreinick <heschi@google.com>
Reviewed-by: Bryan Mills <bcmills@google.com>
TryBot-Result: Gopher Robot <gobot@golang.org>