x/text: bidi incorrect brackets matching
Brackets are now ordered correctly in the bidi algorithm. While parsing the input, the pairValues slice is created with correct bracket data.

The algorithm in matchOpener from brackets.go checks if the opening and closing values are equal. Therefore the unicode value of the closing bracket needs to be changed to the opening bracket.